[FEAT] 메인페이지 및 지역(핫스팟)선택 페이지 ui 연동

+++ src/main/java/com/example/whatseoul/entity/PopulationForecast.java
@@ -1,5 +1,4 @@
 package com.example.whatseoul.entity;
-
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,11 +12,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class PopulationForecast {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "FORECAST_ID")
-    private int forecastId;
+    private Long forecastId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "POPULATION_ID")
